@@ -1,6 +1,6 @@
 {
   "name": "nap",
-  "version": "0.14.1",
+  "version": "0.15.0",
   "main": "index.js",
   "scripts": {
     "setup": "echo \"# See .env.example for more\" > .env2",
@@ -32,7 +32,7 @@
     "cover": "jest --coverage --no-cache"
   },
   "dependencies": {
-    "apollo-client": "^1.2.0",
+    "apollo-client": "^1.2.2",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.17.1",
     "connect-redis": "^3.3.0",
@@ -51,17 +51,17 @@
     "graphql-compose-relay": "^1.0.11",
     "graphql-tag": "^2.0.0",
     "graphql-tools": "^0.11.0",
-    "helmet": "^3.5.0",
+    "helmet": "^3.6.0",
     "isomorphic-fetch": "^2.2.1",
     "jsonwebtoken": "^7.4.0",
+    "localforage": "^1.5.0",
     "lusca": "^1.4.1",
     "mailgun.js": "^2.0.1",
-    "mongoose": "^4.9.7",
+    "mongoose": "^4.9.8",
     "mongoose-role": "^2.0.0",
-    "morgan": "^1.8.1",
     "mubsub": "^1.4.0",
     "multer": "^1.3.0",
-    "next": "^2.2.0",
+    "next": "2.3.1",
     "passport": "^0.3.2",
     "passport-facebook": "^2.1.1",
     "passport-facebook-token": "^3.3.0",
@@ -87,7 +87,7 @@
   "devDependencies": {
     "babel-cli": "^6.24.1",
     "babel-eslint": "^7.2.3",
-    "babel-jest": "^19.0.0",
+    "babel-jest": "^20.0.0",
     "babel-polyfill": "^6.23.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
@@ -100,11 +100,12 @@
     "eslint": "^3.19.0",
     "eslint-plugin-babel": "^4.1.1",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-react": "^6.10.3",
-    "jest": "^19.0.2",
-    "jest-cli": "^19.0.2",
+    "eslint-plugin-react": "^7.0.0",
+    "jest": "^20.0.0",
+    "jest-cli": "^20.0.0",
     "nodemon": "^1.11.0",
     "promise-sequential": "^1.1.1",
+    "prop-types": "^15.5.8",
     "react-addons-test-utils": "^15.5.1",
     "react-test-renderer": "^15.5.4"
   }
