@@ -3,18 +3,20 @@ import React from 'react'
 import 'isomorphic-fetch'
 import { initClient } from './initClient'
 import { initStore } from './initStore'
+import PropTypes from 'prop-types'
 
 export default (Component) => (
   class extends React.Component {
     static propTypes = () => ({
-      headers: React.PropTypes.object.isRequired,
-      graphql_url: React.PropTypes.object.isRequired,
-      initialState: React.PropTypes.object.isRequired
+      headers: PropTypes.object.isRequired,
+      graphql_url: PropTypes.object.isRequired,
+      accessToken: PropTypes.string,
+      initialState: PropTypes.object.isRequired
     })
 
     static async getInitialProps(ctx) {
       const headers = ctx.req ? ctx.req.headers : {}
-      const graphql_url = `${ctx.req.protocol}://${ctx.req.headers.host}/graphql`
+      const graphql_url = ctx.req ? `${ctx.req.protocol}://${ctx.req.headers.host}/graphql` : '/graphql'
       const client = initClient(headers, graphql_url)
       const store = initStore(client, client.initialState)
 
