@@ -1,11 +1,12 @@
 import React from 'react'
 import Router from 'next/router'
 import 'isomorphic-fetch'
+import PropTypes from 'prop-types'
 
 const resetPassword = (token, password) => fetch('/reset-password-by-token', {
   method: 'POST',
   body: JSON.stringify({ token, password }),
-  headers : new Headers({
+  headers: new Headers({
     'Content-Type': 'application/json',
     'Accept': 'application/json'
   })
@@ -30,14 +31,14 @@ class Reset extends React.Component {
       if (json.data.isReset) {
         return Router.push('/auth/reset-succeed')
       }
-    }).catch(err => console.error(err))
+    }).catch(err => console.error(err)) // eslint-disable-line
   }
 
   render() {
     return <form onSubmit={this.handleSubmit}>
       <h1>Reset Password</h1>
       <input placeholder='token' name='token' defaultValue={this.props.token} /><br />
-      <input placeholder='password' name='password' defaultValue='barbar' />
+      <input placeholder='password' name='password' defaultValue='foobar' />
       <button type='submit'>Reset</button>
       <style jsx>{`
       form {
@@ -58,7 +59,7 @@ class Reset extends React.Component {
 }
 
 Reset.propTypes = () => ({
-  resetPassword: React.PropTypes.func.isRequired
+  resetPassword: PropTypes.func.isRequired
 })
 
 export default Reset
